+++ packages/vsce/src/extension.ts
@@ -74,6 +74,8 @@ export async function activate(context: ExtensionContext): Promise<IExtensionAPI
     canSelectMany: true,
   });
 
+  treeDataProv.hookCollapseWatcher(treeview);
+
   const contextMap: { [key: string]: (node: any) => Promise<void> | void } = {
     cicssession: async (node: any) => {
       await sessionExpansionHandler(node.element, treeDataProv);
@@ -103,8 +105,9 @@ export async function activate(context: ExtensionContext): Promise<IExtensionAPI
     if (initialContext in contextMap) {
       contextMap[initialContext](node);
     }
-    // @ts-ignore
-    node.element.refreshIcon(true);
+    if (node.element.refreshIcon) {
+      node.element.refreshIcon(true);
+    }
     treeDataProv._onDidChangeTreeData.fire(node.element);
   });
 
@@ -120,7 +123,7 @@ export async function activate(context: ExtensionContext): Promise<IExtensionAPI
 
   context.subscriptions.push(...getCommands(treeDataProv, treeview, context));
   context.subscriptions.push(
-    window.registerWebviewViewProvider(ResourceInspectorViewProvider.viewType, ResourceInspectorViewProvider.getInstance(context))
+    window.registerWebviewViewProvider(ResourceInspectorViewProvider.viewType, ResourceInspectorViewProvider.getInstance(context, treeDataProv))
   );
 
   return CICSExtenderApiConfig.getAPI();